requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# NLP and ML
transformers>=4.30.0
datasets>=2.12.0
torch>=2.0.0
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
finbert-embedding>=0.1.0

# Financial data and backtesting
yfinance>=0.2.0
vectorbt>=0.25.0
backtrader>=1.9.0
zipline-reloaded>=2.2.0

# Model interpretability
shap>=0.42.0
lime>=0.2.0
eli5>=0.13.0

# Evaluation and metrics
optuna>=3.2.0
mlflow>=2.5.0
wandb>=0.15.0

# Web app
streamlit>=1.25.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: GPU acceleration
# torch-audio>=2.0.0  # Uncomment if using audio features
# accelerate>=0.20.0  # Uncomment for distributed training
